% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnalyzeEntitiesRequest}
\alias{AnalyzeEntitiesRequest}
\title{AnalyzeEntitiesRequest Object}
\usage{
AnalyzeEntitiesRequest(document = NULL, encodingType = NULL)
}
\arguments{
\item{document}{Input document}

\item{encodingType}{The encoding type used by the API to calculate offsets}
}
\value{
AnalyzeEntitiesRequest object
}
\description{
AnalyzeEntitiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The entity analysis request message.
}
\seealso{
Other AnalyzeEntitiesRequest functions: \code{\link{documents.analyzeEntities}}
}

