% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnalyzeSentimentRequest}
\alias{AnalyzeSentimentRequest}
\title{AnalyzeSentimentRequest Object}
\usage{
AnalyzeSentimentRequest(document = NULL)
}
\arguments{
\item{document}{Input document}
}
\value{
AnalyzeSentimentRequest object
}
\description{
AnalyzeSentimentRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The sentiment analysis request message.
}
\seealso{
Other AnalyzeSentimentRequest functions: \code{\link{documents.analyzeSentiment}}
}

