% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnalyzeSentimentResponse}
\alias{AnalyzeSentimentResponse}
\title{AnalyzeSentimentResponse Object}
\usage{
AnalyzeSentimentResponse(documentSentiment = NULL, language = NULL)
}
\arguments{
\item{documentSentiment}{The overall sentiment of the input document}

\item{language}{The language of the text, which will be the same as the language specified}
}
\value{
AnalyzeSentimentResponse object
}
\description{
AnalyzeSentimentResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The sentiment analysis response message.
}

