% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnalyzeSyntaxRequest}
\alias{AnalyzeSyntaxRequest}
\title{AnalyzeSyntaxRequest Object}
\usage{
AnalyzeSyntaxRequest(encodingType = NULL, document = NULL)
}
\arguments{
\item{encodingType}{The encoding type used by the API to calculate offsets}

\item{document}{Input document}
}
\value{
AnalyzeSyntaxRequest object
}
\description{
AnalyzeSyntaxRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The syntax analysis request message.
}
\seealso{
Other AnalyzeSyntaxRequest functions: \code{\link{documents.analyzeSyntax}}
}
