% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnalyzeSyntaxResponse}
\alias{AnalyzeSyntaxResponse}
\title{AnalyzeSyntaxResponse Object}
\usage{
AnalyzeSyntaxResponse(language = NULL, sentences = NULL, tokens = NULL)
}
\arguments{
\item{language}{The language of the text, which will be the same as the language specified}

\item{sentences}{Sentences in the input document}

\item{tokens}{Tokens, along with their syntactic information, in the input document}
}
\value{
AnalyzeSyntaxResponse object
}
\description{
AnalyzeSyntaxResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The syntax analysis response message.
}
