% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{DependencyEdge}
\alias{DependencyEdge}
\title{DependencyEdge Object}
\usage{
DependencyEdge(headTokenIndex = NULL, label = NULL)
}
\arguments{
\item{headTokenIndex}{Represents the head of this token in the dependency tree}

\item{label}{The parse label for the token}
}
\value{
DependencyEdge object
}
\description{
DependencyEdge Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents dependency parse tree information for a token.
}

