% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Document}
\alias{Document}
\title{Google Cloud Natural Language API Objects 
Google Cloud Natural Language API provides natural language understanding technologies to developers. Examples include sentiment analysis, entity recognition, and text annotations.}
\usage{
Document(language = NULL, gcsContentUri = NULL, type = NULL,
  content = NULL)
}
\arguments{
\item{language}{The language of the document (if not specified, the language is}

\item{gcsContentUri}{The Google Cloud Storage URI where the file content is located}

\item{type}{Required}

\item{content}{The content of the input in string format}
}
\value{
Document object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:23:53
filename: /Users/mark/dev/R/autoGoogleAPI/googlelanguagev1beta1.auto/R/language_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Document Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
################################################################ #Represents the input to API methods.
}

