% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_functions.R
\name{documents.analyzeSentiment}
\alias{documents.analyzeSentiment}
\title{Analyzes the sentiment of the provided text.}
\usage{
documents.analyzeSentiment(AnalyzeSentimentRequest)
}
\arguments{
\item{AnalyzeSentimentRequest}{The \link{AnalyzeSentimentRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/natural-language/}{Google Documentation}

Other AnalyzeSentimentRequest functions: \code{\link{AnalyzeSentimentRequest}}
}

