% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Entity.metadata}
\alias{Entity.metadata}
\title{Entity.metadata Object}
\usage{
Entity.metadata()
}
\value{
Entity.metadata object
}
\description{
Entity.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata associated with the entity.Currently, only Wikipedia URLs are provided, if available.The associated key is 'wikipedia_url'.
}
\seealso{
Other Entity functions: \code{\link{Entity}}
}

