% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Entity}
\alias{Entity}
\title{Entity Object}
\usage{
Entity(Entity.metadata = NULL, metadata = NULL, salience = NULL,
  type = NULL, mentions = NULL, name = NULL)
}
\arguments{
\item{Entity.metadata}{The \link{Entity.metadata} object or list of objects}

\item{metadata}{Metadata associated with the entity}

\item{salience}{The salience score associated with the entity in the [0, 1}

\item{type}{The entity type}

\item{mentions}{The mentions of this entity in the input document}

\item{name}{The representative name for the entity}
}
\value{
Entity object
}
\description{
Entity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a phrase in the text that is a known entity, such asa person, an organization, or location. The API associates information, suchas salience and mentions, with entities.
}
\seealso{
Other Entity functions: \code{\link{Entity.metadata}}
}

