% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{PartOfSpeech}
\alias{PartOfSpeech}
\title{PartOfSpeech Object}
\usage{
PartOfSpeech(form = NULL, number = NULL, voice = NULL, aspect = NULL,
  mood = NULL, tag = NULL, gender = NULL, person = NULL,
  proper = NULL, case = NULL, tense = NULL, reciprocity = NULL)
}
\arguments{
\item{form}{The grammatical form}

\item{number}{The grammatical number}

\item{voice}{The grammatical voice}

\item{aspect}{The grammatical aspect}

\item{mood}{The grammatical mood}

\item{tag}{The part of speech tag}

\item{gender}{The grammatical gender}

\item{person}{The grammatical person}

\item{proper}{The grammatical properness}

\item{case}{The grammatical case}

\item{tense}{The grammatical tense}

\item{reciprocity}{The grammatical reciprocity}
}
\value{
PartOfSpeech object
}
\description{
PartOfSpeech Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents part of speech information for a token.
}
