% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Sentiment}
\alias{Sentiment}
\title{Sentiment Object}
\usage{
Sentiment(polarity = NULL, magnitude = NULL)
}
\arguments{
\item{polarity}{Polarity of the sentiment in the [-1}

\item{magnitude}{A non-negative number in the [0, +inf) range, which represents}
}
\value{
Sentiment object
}
\description{
Sentiment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the feeling associated with the entire text or entities inthe text.
}

