% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Entity}
\alias{Entity}
\title{Google Cloud Natural Language API Objects 
Google Cloud Natural Language API provides natural language understanding technologies to developers. Examples include sentiment analysis, entity recognition, and text annotations.}
\usage{
Entity(Entity.metadata = NULL, mentions = NULL, name = NULL,
  type = NULL, metadata = NULL, salience = NULL)
}
\arguments{
\item{Entity.metadata}{The \link{Entity.metadata} object or list of objects}

\item{mentions}{The mentions of this entity in the input document}

\item{name}{The representative name for the entity}

\item{type}{The entity type}

\item{metadata}{Metadata associated with the entity}

\item{salience}{The salience score associated with the entity in the [0, 1}
}
\value{
Entity object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:00:25
filename: /Users/mark/dev/R/autoGoogleAPI/googlelanguagev1beta1.auto/R/language_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Entity Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a phrase in the text that is a known entity, such asa person, an organization, or location. The API associates information, suchas salience and mentions, with entities.
}
\seealso{
Other Entity functions: \code{\link{Entity.metadata}}
}
