% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{PartOfSpeech}
\alias{PartOfSpeech}
\title{PartOfSpeech Object}
\usage{
PartOfSpeech(tag = NULL)
}
\arguments{
\item{tag}{The part of speech tag}
}
\value{
PartOfSpeech object
}
\description{
PartOfSpeech Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents part of speech information for a token.
}

